+++ src/array/range.pipe.ts
@@ -6,11 +6,12 @@ import { isArray } from '../utils/utils';
 })
 export class RangePipe implements PipeTransform {
     
-    transform (input: any, [size = 0, start = 1]): any {
+    transform (input: any, [size = 0, start = 1, step = 1]): any {
         
         const range: number[] = [];
         for (let length = 0; length < size; ++length) {
-            range.push(start++);
+            range.push(start);
+            start += step;
         }
         
         return range;